Tweak signature of the matrix_loop() function for use as an interrupt handler.
The matrix_loop() function was observed to cause instability at startup due to
races between python's module loading and the timer ISR (you would occasionally
get a NameEerror and the timer would die). To get around this, we change the
matrix_loop() to accept any arguments, which are all ignored, so that the ISR
can call it directly without needing to do any name resolution. This seems to
resolve the problem.